+++ CHANGELOG.md
@@ -6,6 +6,8 @@
  - Category export only create/update categories from the exported channel
  - Make it compatible with EE-1.11 and EE-1.12
  - Add a log to profile calls to Magento
+ - Add a command to purge Mapping in database
+ - Add a command to purge Delta in database
 
 ## BC Breaks
  - All DeltaExportBundle dependencies should be replaced by MagentoConnectorBundle ones